Major update
Changed many aspects of the button. Notably, another gesture to it to allow addition interaction in scan history page as well as many small errors that were in the component recently. The name of the component file should match the name  of the component in the file. The component was also missing an export statement on its declaration. Just needs UI/UX so it does not look so barebones.